refactor: extract edit-collection.js
